EEQ-BC Implementation (#56)
This PR contains the implementation of the EEQ-BC charge model (see
[here](https://doi.org/10.1063/5.0268978)).

The implementation contains full support for
- electrostatic energy with gradients and stress tensor
- atomic partial charges with gradients and lattice vector derivatives
- both molecular and periodic calculations

EEQ-BC is implemented as an alternative model. The general solution of
the linear system remains unchanged. The new capacitance matrix is
handled in the cache and modifies only the final A-matrix and b-vector.

Additionally, we added explicit routines to calculate the EEQ and EEQ-BC
charges, avoiding the through DFT-D4.

---------

Co-authored-by: Polt <seidler118@gmail.com>
Co-authored-by: Leopold Seidler <100208101+lmseidler@users.noreply.github.com>
Co-authored-by: Marvin Friede <51965259+marvinfriede@users.noreply.github.com>